+++ Data_Buffer/data_buffer.h
@@ -9,7 +9,7 @@ typedef struct Buffer
 {
     int capacity;
     int elementsInQueue;
-    int* queue;
+    double* queue;
 } Buffer;
 
 
@@ -19,11 +19,11 @@ Buffer CreateNewBuffer(int bufferSize);
 
 
 //this function adds a new data point to the end of the buffer queue, removing the first data point in the queue if the buffer is full
-void Enqueue(Buffer* b, int dataPoint);
+void Enqueue(Buffer* b, double dataPoint);
 
 
 //this function returns the first data point in the buffer and then removes it from the buffer, assuming the buffer is not empty
-//returns the max value for int if the buffer is empty as an error indicator
-int Dequeue(Buffer* b);
+//returns -1 if the buffer is empty as an error indicator
+double Dequeue(Buffer* b);
 
 #endif
